update replace_result_with_data!
made it a little more explicit that score is also an optional param,
since canvas accepts either a score or return data.  If nothing is
passed the XML will still build correctly, however, the tool consumer
will likely reject the submission since it would mean that there is
nothing to submit.